# api/src/main/java/com/kipcollo/products/ProductCategoryRequest.java
package com.kipcollo.products;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.List;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class ProductCategoryRequest {

    private String name;
    private String description;
    private byte[] image;
}