# pharmacy-ui/src/app/services/models/order-line.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Orders } from '../models/orders';
export interface OrderLine {
  orderLineId?: number;
  orders?: Orders;
  productId?: number;
  quantity?: number;
  status?: 'PROCESSING' | 'SHIPPED' | 'DELIVERED';
}