# pharmacy-ui/src/app/services/models/order-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Product } from '../models/product';
import { Users } from '../models/users';
export interface OrderResponse {
  createdAt?: string;
  customers?: Users;
  lastModifiedDate?: string;
  localDateTime?: string;
  orderId?: number;
  paymentMethod?: 'MPESA' | 'BANK' | 'BITCOIN' | 'VISA';
  products?: Array<Product>;
  reference?: string;
  totalAmount?: number;
}