# pharmacy-ui/src/app/services/models/orders.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { OrderLine } from '../models/order-line';
import { Product } from '../models/product';
import { Users } from '../models/users';
export interface Orders {
  createdAt?: string;
  customerId?: number;
  customers?: Users;
  id?: number;
  lastModifiedDate?: string;
  localDateTime?: string;
  orderLines?: Array<OrderLine>;
  paymentMethod?: 'MPESA' | 'BANK' | 'BITCOIN' | 'VISA';
  prescriptionId?: number;
  products?: Array<Product>;
  reference?: string;
  totalAmount?: number;
}