# pharmacy-ui/src/app/services/models/product-category-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Product } from '../models/product';
export interface ProductCategoryResponse {
  description?: string;
  id?: number;
  image?: string;
  name?: string;
  products?: Array<Product>;
}