# pharmacy-ui/src/app/services/models/product.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProductCategory } from '../models/product-category';
export interface Product {
  category?: ProductCategory;
  createdDate?: string;
  description?: string;
  expiryDate?: string;
  id?: number;
  image?: string;
  lastModifiedDate?: string;
  manufacturer?: string;
  name?: string;
  price?: number;
  stockQuantity?: number;
  type?: 'SYRUP' | 'TABLETS' | 'OINTMENTS';
}