# pharmacy-ui/src/app/services/models/users.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { GrantedAuthority } from '../models/granted-authority';
import { Roles } from '../models/roles';
import { Token } from '../models/token';
export interface Users {
  accountLocked?: boolean;
  accountNonExpired?: boolean;
  accountNonLocked?: boolean;
  authorities?: Array<GrantedAuthority>;
  createdDate?: string;
  credentialsNonExpired?: boolean;
  customerId?: number;
  dateOfBirthDate?: string;
  email?: string;
  enabled?: boolean;
  firstName?: string;
  lastModifiedDate?: string;
  lastName?: string;
  location?: string;
  name?: string;
  password?: string;
  phone?: string;
  profilePicture?: string;
  roles?: Array<Roles>;
  tokens?: Array<Token>;
  username?: string;
}