# pharmacy-ui/src/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/products":{"get":{"tags":["Medicine APIs"],"operationId":"getAllMedicines","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseProductResponse"}}}}}},"put":{"tags":["Medicine APIs"],"operationId":"updateMedicine","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"medicine":{"$ref":"#/components/schemas/ProductRequest"},"image":{"type":"string","format":"binary"}}}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["Medicine APIs"],"operationId":"createMedicine","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"product":{"$ref":"#/components/schemas/ProductRequest"},"image":{"type":"string","format":"binary"}},"required":["image","product"]}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/prescriptions/{id}/approve":{"put":{"tags":["prescription-controller"],"operationId":"approvePrescription","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/customers":{"get":{"tags":["Customers APIs"],"operationId":"getAllCustomers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"put":{"tags":["Customers APIs"],"operationId":"updateCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["Customers APIs"],"operationId":"addCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/auth/register":{"post":{"tags":["Authentication APIs"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication APIs"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/products/purchase":{"post":{"tags":["Medicine APIs"],"operationId":"purchaseProducts","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PurchaseProductRequest"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PurchaseProductResponse"}}}}}}}},"/api/product-category":{"get":{"tags":["product-category-controller"],"operationId":"getProductCategory","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductCategoryResponse"}}}}}}},"post":{"tags":["product-category-controller"],"operationId":"createProductCategory","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"product":{"$ref":"#/components/schemas/ProductCategoryRequest"},"image":{"type":"string","format":"binary"}},"required":["image","product"]}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/prescriptions":{"get":{"tags":["prescription-controller"],"operationId":"getAllPrescriptions","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PrescriptionResponse"}}}}}}},"post":{"tags":["prescription-controller"],"operationId":"uploadPrescription","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/payment":{"post":{"tags":["payment-controller"],"operationId":"createPayment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/api/orders":{"get":{"tags":["Order APIs"],"operationId":"findAll","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderResponse"}}}}}}},"post":{"tags":["Order APIs"],"operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/api/orders/create":{"post":{"tags":["Order APIs"],"operationId":"createPrescriptionOrder","parameters":[{"name":"prescriptionId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/cart":{"post":{"tags":["cart-controller"],"operationId":"addCart","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartResponse"}}}}}}},"/api/cart/checkout/{userId}":{"post":{"tags":["cart-controller"],"operationId":"placeOrder","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/auth/activate-account":{"get":{"tags":["Authentication APIs"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/products/{id}":{"get":{"tags":["Medicine APIs"],"operationId":"getMedicineById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"delete":{"tags":["Medicine APIs"],"operationId":"deleteMedicine","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/product-category/{id}":{"get":{"tags":["product-category-controller"],"operationId":"getProductCategoryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCategoryResponse"}}}}}}},"/api/orders/{orderId}":{"get":{"tags":["Order APIs"],"operationId":"findById","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}}}},"/api/orders/{id}/status":{"get":{"tags":["Order APIs"],"operationId":"getOrderStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/orderline/order/{orderId}":{"get":{"tags":["order-line-controller"],"operationId":"findOrderById","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderLineResponse"}}}}}}}},"/api/customers/{customerId}":{"get":{"tags":["Customers APIs"],"operationId":"getCustomer","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"delete":{"tags":["Customers APIs"],"operationId":"deleteCustomer","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/customers/exists/{customerId}":{"get":{"tags":["Customers APIs"],"operationId":"existsById","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"boolean"}}}}}}},"/api/customers/currentCustomer":{"get":{"tags":["Customers APIs"],"operationId":"getCurrentCustomer","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/cart/status/{status}":{"get":{"tags":["cart-controller"],"operationId":"getCartsByStatus","parameters":[{"name":"status","in":"path","required":true,"schema":{"type":"string","enum":["STARTED","CHECKED_OUT","ABANDONED","REMOVED"]}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CartResponse"}}}}}}}},"/api/cart/add/":{"get":{"tags":["cart-controller"],"operationId":"getUserCart","parameters":[{"name":"securityContext","in":"query","required":true,"schema":{"$ref":"#/components/schemas/SecurityContext"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CartResponse"}}}}}}}},"/api/cart/{cartId}":{"delete":{"tags":["cart-controller"],"operationId":"removeFromCart","parameters":[{"name":"cartId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"ProductRequest":{"type":"object","properties":{"medicineId":{"type":"integer","format":"int32"},"name":{"type":"string"},"image":{"type":"string","format":"byte"},"description":{"type":"string"},"type":{"type":"string","enum":["SYRUP","TABLETS","OINTMENTS"]},"manufacturer":{"type":"string"},"expiryDate":{"type":"string","format":"date"},"stockQuantity":{"type":"integer","format":"int32"},"price":{"type":"number"}}},"UserRequest":{"type":"object","properties":{"customerId":{"type":"integer","format":"int32"},"firstName":{"type":"string"},"lastName":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"phone":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"location":{"type":"string"}},"required":["email","firstName","lastName","password"]},"RegistrationRequest":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["email","firstName","lastName","password"]},"AuthenticationRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["email","password"]},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}},"PurchaseProductRequest":{"type":"object","properties":{"productId":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"}},"required":["productId","quantity"]},"PurchaseProductResponse":{"type":"object","properties":{"productId":{"type":"integer","format":"int32"},"name":{"type":"string"},"image":{"type":"string","format":"byte"},"description":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"number","format":"double"}}},"ProductCategoryRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"image":{"type":"string","format":"byte"}}},"PaymentRequest":{"type":"object","properties":{"paymentId":{"type":"integer","format":"int32"},"amount":{"type":"number"},"paymentMethod":{"type":"string","enum":["MPESA","BANK","BITCOIN","VISA"]},"orderId":{"type":"integer","format":"int32"},"orderReference":{"type":"string"},"customer":{"$ref":"#/components/schemas/UserResponse"}}},"Roles":{"type":"object","properties":{"roleId":{"type":"integer","format":"int32"},"name":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"UserResponse":{"type":"object","properties":{"customerId":{"type":"integer","format":"int32"},"firstName":{"type":"string"},"lastName":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"phone":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"location":{"type":"string"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/Roles"}},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"OrderLine":{"type":"object","properties":{"orderLineId":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["PROCESSING","SHIPPED","DELIVERED"]},"orders":{"$ref":"#/components/schemas/Orders"},"productId":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"}}},"OrderRequest":{"type":"object","properties":{"order":{"$ref":"#/components/schemas/Orders"},"orderId":{"type":"integer","format":"int32"},"reference":{"type":"string"},"paymentMethod":{"type":"string","enum":["MPESA","BANK","BITCOIN","VISA"]},"customers":{"$ref":"#/components/schemas/Users"},"totalAmount":{"type":"number"},"products":{"type":"array","items":{"$ref":"#/components/schemas/PurchaseProductRequest"}}},"required":["customers","paymentMethod","products","reference"]},"Orders":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"reference":{"type":"string"},"totalAmount":{"type":"number"},"paymentMethod":{"type":"string","enum":["MPESA","BANK","BITCOIN","VISA"]},"customers":{"$ref":"#/components/schemas/Users"},"products":{"type":"array","items":{"$ref":"#/components/schemas/Product"}},"localDateTime":{"type":"string","format":"date-time"},"orderLines":{"type":"array","items":{"$ref":"#/components/schemas/OrderLine"}},"prescriptionId":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"},"customerId":{"type":"integer","format":"int32"}}},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"type":{"type":"string","enum":["SYRUP","TABLETS","OINTMENTS"]},"image":{"type":"string","format":"byte"},"manufacturer":{"type":"string"},"expiryDate":{"type":"string","format":"date"},"stockQuantity":{"type":"integer","format":"int32"},"price":{"type":"number"},"category":{"$ref":"#/components/schemas/ProductCategory"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"ProductCategory":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"image":{"type":"string","format":"byte"},"products":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}},"Token":{"type":"object","properties":{"tokenId":{"type":"integer","format":"int32"},"token":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"expiresAt":{"type":"string","format":"date-time"},"validatedAt":{"type":"string","format":"date-time"},"users":{"$ref":"#/components/schemas/Users"}}},"Users":{"type":"object","properties":{"customerId":{"type":"integer","format":"int32"},"firstName":{"type":"string"},"lastName":{"type":"string"},"dateOfBirthDate":{"type":"string","format":"date"},"profilePicture":{"type":"string","format":"byte"},"email":{"type":"string"},"password":{"type":"string"},"phone":{"type":"string"},"location":{"type":"string"},"accountLocked":{"type":"boolean"},"enabled":{"type":"boolean"},"tokens":{"type":"array","items":{"$ref":"#/components/schemas/Token"}},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/Roles"}},"username":{"type":"string"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"name":{"type":"string"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"}}},"CartRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"},"product":{"type":"array","items":{"$ref":"#/components/schemas/Product"}},"ordered":{"type":"boolean"}}},"CartResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"},"product":{"type":"array","items":{"$ref":"#/components/schemas/Product"}},"ordered":{"type":"boolean"},"status":{"type":"string","enum":["STARTED","CHECKED_OUT","ABANDONED","REMOVED"]}}},"PageResponseProductResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"ProductResponse":{"type":"object","properties":{"productId":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"type":{"type":"string","enum":["SYRUP","TABLETS","OINTMENTS"]},"image":{"type":"string","format":"byte"},"manufacturer":{"type":"string"},"expiryDate":{"type":"string","format":"date"},"stockQuantity":{"type":"integer","format":"int32"},"price":{"type":"number"}}},"ProductCategoryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"image":{"type":"string","format":"byte"},"products":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}},"PrescriptionResponse":{"type":"object"},"OrderResponse":{"type":"object","properties":{"orderId":{"type":"integer","format":"int32"},"reference":{"type":"string"},"customers":{"$ref":"#/components/schemas/Users"},"products":{"type":"array","items":{"$ref":"#/components/schemas/Product"}},"paymentMethod":{"type":"string","enum":["MPESA","BANK","BITCOIN","VISA"]},"totalAmount":{"type":"number"},"localDateTime":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"OrderLineResponse":{"type":"object","properties":{"orderLineId":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"}}},"SecurityContext":{"type":"object"}}}}